client/requirements_client.txt
pillow
pygame
websocket-client
